+++ cdk/.flake8
@@ -2,3 +2,4 @@
 max-complexity = 15
 # Default plus E266
 ignore = E121,E123,E126,E226,E24,E704,W503,W504,E266,E203,E501
+per-file-ignores = __init__.py:F401
