+++ cdk/tests/unit/config/__init__.py
@@ -1,14 +1,7 @@
 import unittest
 from unittest.mock import patch
 
-from domino_cdk.config import config_loader, config_template
-from domino_cdk.config.base import DominoCDKConfig
-from domino_cdk.config.efs import EFS
-from domino_cdk.config.eks import EKS
-from domino_cdk.config.route53 import Route53
-from domino_cdk.config.s3 import S3
-from domino_cdk.config.util import IngressRule
-from domino_cdk.config.vpc import VPC
+from domino_cdk.config import config_loader, config_template, DominoCDKConfig, EFS, EKS, Route53, S3, IngressRule, VPC
 
 default_config = DominoCDKConfig(
     name='domino',
