+++ cdk/tests/unit/config/test_util.py
@@ -2,7 +2,8 @@ import unittest
 from copy import deepcopy
 from unittest.mock import patch
 
-from domino_cdk.config.util import IngressRule, check_leavins, from_loader
+from domino_cdk.config import IngressRule
+from domino_cdk.config.util import check_leavins, from_loader
 
 ingress_rules = [{"name": "some_rule", "from_port": 22, "to_port": 22, "protocol": "TCP", "ip_cidrs": ["10.0.0.0/16"]}]
